[refactor] Text.js moved into new folder display
